Update selection after executing operations rather than before
The execution of a transformation function or a handler is always handled by
the Surface instance now, which guarantees that the selection is always up to
date after.

This prevents some bugs where the selection wouldn't be up to date when
executing an operation, thus calling the wrong handler or worse, allowing
content-editable's default behavior. (Some of them were already fixed when
the selection was updated before the execution, but that should fix all of
them and fit more nicely with things like toolbars, etc.)