Add selection save/restore to the command history
The state of selection just before a command executes is now saved with the
command itself and is restored by the History instance.
The implementation is quick and simple but feels a bit hacky... Perhaps it
would be better (and more logical) to move all model modification methods
upwards to the Editor instance itself, that would serve as context for event
handlers instead of Surface instances.